Update standards, remove unfinished `als`, + more
Changes:

  * `als` is too big a project for me right now, so it's gone, for now.
  * Updated files to the standard 80-column limit.
  * Addressed issue when `libtflbp-sh` files may be missing.
  * Prefer `#!/usr/bin/env bash`, for portability.
  * Made me use `libtflbp-sh` much more, in older files.
  * Tidied up some of the wording in various `Usage()` outputs.
  * Added list of dependencies to the header of some files.
  * Made `links.sh` operate more contextually.
  * Removed or disabled some unused or unneeded code.

+++ devutils/links.sh
@@ -1,34 +1,46 @@
 #!/bin/sh
 
-#----------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
 # Project Name      - Extra/devutils/links.sh
 # Started On        - Sun 22 Oct 00:15:02 BST 2017
-# Last Change       - Mon  9 Dec 18:45:57 GMT 2019
+# Last Change       - Fri 31 Jan 02:47:34 GMT 2020
 # Author E-Mail     - terminalforlife@yahoo.com
 # Author GitHub     - https://github.com/terminalforlife
-#----------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
 # Just a simple, quick script to update the hard links when changing branches.
-#------------------------------------------------------------------------------MAIN
+#------------------------------------------------------------------------------
 
+set -e
 . /usr/lib/tflbp-sh/ChkDep
+set +e
 
 ChkDep ln rm
 
 {
 	cd "$HOME/GitHub/terminalforlife/Personal/Extra"
 
-	rm -v $HOME/.config/compton.conf
-	ln -v misc/compton.conf $HOME/.config/
+	if command -v compton 1> /dev/null; then
+		rm -v $HOME/.config/compton.conf
+		ln -v misc/compton.conf $HOME/.config/
+	fi
 
-	rm -v $HOME/.keynavrc
-	ln -v misc/.keynavrc $HOME/
+	if command -v keynav 1> /dev/null; then
+		rm -v $HOME/.keynavrc
+		ln -v misc/.keynavrc $HOME/
+	fi
 
-	rm -v $HOME/.config/dunst/dunstrc
-	ln -v misc/dunstrc $HOME/.config/dunst/
+	if command -v dunst 1> /dev/null; then
+		rm -v $HOME/.config/dunst/dunstrc
+		ln -v misc/dunstrc $HOME/.config/dunst/
+	fi
 
-	rm -v $HOME/.config/xfce4/terminal/terminalrc
-	ln -v misc/terminalrc $HOME/.config/xfce4/terminal/terminalrc
+	if command -v xfce4-terminal 1> /dev/null; then
+		rm -v $HOME/.config/xfce4/terminal/terminalrc
+		ln -v misc/terminalrc $HOME/.config/xfce4/terminal/terminalrc
+	fi
 
-	rm -v $HOME/.config/herbstluftwm/autostart
-	ln -v misc/autostart $HOME/.config/herbstluftwm/
+	if command -v herbsluftwm 1> /dev/null; then
+		rm -v $HOME/.config/herbstluftwm/autostart
+		ln -v misc/autostart $HOME/.config/herbstluftwm/
+	fi
 } 2> /dev/null
